[Refactor] Writing 재정리 + handleTextareaChange 함수 따로 뺐음(#61)
