+++ src/utils/textareaHandler.js
@@ -1,0 +1,8 @@
+const handleTextareaChange = (textareaRef) => {
+  if (textareaRef.current) {
+      textareaRef.current.style.height = 'auto';
+      textareaRef.current.style.height = textareaRef.current.scrollHeight + 'px';
+  }
+};
+
+export { handleTextareaChange }
